Move owid_energy to external channel
